+++ projects/brazil_bu_chem/scripts/pavedroads.R
@@ -1,3 +1,22 @@
+# ressuspensao gera PM e PM10
+language <- "portuguese" # english spanish
+metadata <- readRDS("config/metadata.rds")
+mileage <- readRDS("config/mileage.rds")
+tfs <- readRDS("config/tfs.rds")
+net <- readRDS("network/net.rds")
+veh <- readRDS("config/fleet_age.rds")
+lkm <- net$lkm
+tf_PC <- tfs$PC_G
+tf_LCV <- tfs$LCV_G
+tf_TRUCKS <- tfs$TRUCKS_L_D
+tf_BUS <- tfs$BUS_URBAN_D
+tf_MC <- tfs$MC_150_G
+sL1 <- 2.4 # silt [g/m^2] se ADT < 500 (CENMA CHILE) i
+sL2 <- 0.7 # silt [g/m^2] se 500 < ADT < 5000 (CENMA CHILE)
+sL3 <- 0.6 # silt [g/m^2] se 5000 < ADT < 10000 (CENMA CHILE)
+sL4 <- 0.3 # silt [g/m^2] se ADT > 10000 (CENMA CHILE)
+
+
 # identicar nomes de grupos
 nveh <- names(veh)
 n_PC <- nveh[grep(pattern = "PC", x = nveh)]