+++ NEWS.md
@@ -6,6 +6,7 @@ NEWS
 - add more env chinese ef
 - remove ive. It is better to run MOVES and extract EF
 - add project china_bu_chem
+- add more vehicles in ef_wear
 
 ### vein 0.9.91